Fix | CheckUpgradeAndRemoveAsync
