+++ ShadowPluginLoader.WinUI/Interfaces/IPluginLoader.Change.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+
+namespace ShadowPluginLoader.WinUI.Interfaces;
+
+public partial interface IPluginLoader<TMeta, TAPlugin>
+{
+
+    /// <summary>
+    /// Checked for updates and removed plugins
+    /// </summary>
+    /// <returns></returns>
+    Task CheckUpgradeAndRemoveAsync();
+}
