+++ ShadowPluginLoader.WinUI/Interfaces/IPluginLoader.Install.cs
@@ -5,8 +5,6 @@ using System.Threading.Tasks;
 namespace ShadowPluginLoader.WinUI.Interfaces;
 
 public partial interface IPluginLoader<TMeta, TAPlugin>
-    where TAPlugin : AbstractPlugin<TMeta>
-    where TMeta : AbstractPluginMetaData
 {
     /// <summary>
     /// Install From .sdow file
